Extract useConfigurationFolder from the configurations page
